Add dudu package and shift controller, model and utils
