+++ src/main/java/dudu/controller/ByeCommand.java
@@ -1,8 +1,8 @@
-package controller;
+package dudu.controller;
 
-import model.TaskList;
-import utils.DuduException;
-import utils.Ui;
+import dudu.model.TaskList;
+import dudu.utils.DuduException;
+import dudu.utils.Ui;
 
 import java.io.File;
