+++ src/main/java/dudu/controller/Command.java
@@ -1,7 +1,7 @@
-package controller;
+package dudu.controller;
 
-import model.TaskList;
-import utils.DuduException;
+import dudu.model.TaskList;
+import dudu.utils.DuduException;
 
 import java.io.File;
 import java.io.IOException;
