+++ src/main/java/dudu/controller/ListCommand.java
@@ -1,9 +1,9 @@
-package controller;
+package dudu.controller;
 
-import model.Task;
-import model.TaskList;
-import utils.DuduException;
-import utils.Ui;
+import dudu.model.Task;
+import dudu.model.TaskList;
+import dudu.utils.DuduException;
+import dudu.utils.Ui;
 
 import java.io.File;
 import java.util.ArrayList;
