+++ src/main/java/dudu/controller/UnmarkCommand.java
@@ -1,18 +1,17 @@
-package controller;
+package dudu.controller;
 
-import model.Task;
-import model.TaskList;
-import utils.DuduException;
-import utils.Ui;
+import dudu.model.Task;
+import dudu.model.TaskList;
+import dudu.utils.DuduException;
+import dudu.utils.Ui;
 
 import java.io.File;
 import java.io.IOException;
 
-public class MarkCommand implements Command{
-
+public class UnmarkCommand implements Command {
     private final String description;
 
-    public MarkCommand(String description) {
+    public UnmarkCommand(String description) {
         this.description = description;
     }
 
@@ -21,18 +20,20 @@ public class MarkCommand implements Command{
             throw new DuduException("Invalid task number! Please enter a valid task number.");
         }
     }
+
     @Override
     public void execute(TaskList tasks, File cachedTasks) throws DuduException, IOException {
         int index = Integer.parseInt(description) - 1;
         validateIndex(index, tasks);
         Task curr = tasks.get(index);
-        if (!curr.getIsDone()) {
+        if (curr.getIsDone()) {
             curr.toggleDone();
-            String toPrint = "Nice! I've marked this task as done:\n" + curr;
+            String toPrint = "Alright, I've marked this task as not done yet:\n" + curr;
             Ui.printContent(toPrint);
             FileOperation.overwriteFile(cachedTasks, tasks);
         } else {
-            throw new DuduException("Task has already been marked as done!");
+            throw new DuduException("Task is already marked as not done!");
+
         }
     }
