+++ src/main/java/dudu/model/TaskList.java
@@ -1,4 +1,4 @@
-package model;
+package dudu.model;
 
 import java.util.ArrayList;
