+++ src/main/java/dudu/model/Todo.java
@@ -1,4 +1,4 @@
-package model;
+package dudu.model;
 
 public class Todo extends Task {
