+++ src/main/java/dudu/utils/Ui.java
@@ -1,6 +1,4 @@
-package utils;
-
-import model.TaskList;
+package dudu.utils;
 
 import java.util.Scanner;
 
@@ -40,7 +38,7 @@ public class Ui {
         System.out.println("Shows the commands available.");
         printLine();
         System.out.println("bye");
-        System.out.println("Dismisses Dudu.");
+        System.out.println("Dismisses dudu.Dudu.");
         printLine();
     }
 
@@ -58,7 +56,7 @@ public class Ui {
 
     public static void printGreeting() {
         printLine();
-        System.out.println("Hello! I'm Dudu!");
+        System.out.println("Hello! I'm dudu.Dudu!");
         System.out.println("How may I assist you today?");
         printLine();
     }
