+++ text-ui-test/EXPECTED.TXT
@@ -10,44 +10,44 @@ How may I assist you today?
 --------------------------------------------------------
 --------------------------------------------------------
 Alright! I've added this task:
-[T][ ] borrow book
+[T][ ] Yasmeen's Birthday
 Now you have 1 task in the list.
 --------------------------------------------------------
 --------------------------------------------------------
 Alright! I've added this task:
-[D][ ] return book (by: Sunday)
+[D][ ] CS2109S PS1 (by: 1/2/2025 1600)
 Now you have 2 tasks in the list.
 --------------------------------------------------------
 --------------------------------------------------------
 Alright! I've added this task:
-[E][ ] groupwork (from: Thurs 4pm to: 6pm)
+[E][ ] Answering Techniques BBQ (from: 1/2/2025 1800 to: 1/2/2025 2300)
 Now you have 3 tasks in the list.
 --------------------------------------------------------
 --------------------------------------------------------
 Nice! I've marked this task as done:
-[T][X] borrow book
+[T][X] Yasmeen's Birthday
 --------------------------------------------------------
 --------------------------------------------------------
 Nice! I've marked this task as done:
-[E][X] groupwork (from: Thurs 4pm to: 6pm)
+[E][X] Answering Techniques BBQ (from: 1/2/2025 1800 to: 1/2/2025 2300)
 --------------------------------------------------------
 --------------------------------------------------------
-1.[T][X] borrow book
-2.[D][ ] return book (by: Sunday)
-3.[E][X] groupwork (from: Thurs 4pm to: 6pm)
+1.[T][X] Yasmeen's Birthday
+2.[D][ ] CS2109S PS1 (by: 1/2/2025 1600)
+3.[E][X] Answering Techniques BBQ (from: 1/2/2025 1800 to: 1/2/2025 2300)
 --------------------------------------------------------
 --------------------------------------------------------
 Alright, I've marked this task as not done yet:
-[T][ ] borrow book
+[T][ ] Yasmeen's Birthday
 --------------------------------------------------------
 --------------------------------------------------------
 Nice! I've marked this task as done:
-[D][X] return book (by: Sunday)
+[D][X] CS2109S PS1 (by: 1/2/2025 1600)
 --------------------------------------------------------
 --------------------------------------------------------
-1.[T][ ] borrow book
-2.[D][X] return book (by: Sunday)
-3.[E][X] groupwork (from: Thurs 4pm to: 6pm)
+1.[T][ ] Yasmeen's Birthday
+2.[D][X] CS2109S PS1 (by: 1/2/2025 1600)
+3.[E][X] Answering Techniques BBQ (from: 1/2/2025 1800 to: 1/2/2025 2300)
 --------------------------------------------------------
 Invalid task number! Please enter a valid task number.
 Invalid task number! Please enter a valid task number.
